Inclusão de regra de delete

+++ src/main/java/com/acolher/api/service/impl/EnderecoServiceImpl.java
@@ -1,5 +1,6 @@
 package com.acolher.api.service.impl;
 
+import java.security.InvalidParameterException;
 import java.util.List;
 import java.util.Optional;
 
@@ -9,7 +10,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.acolher.api.domain.Endereco;
+import com.acolher.api.domain.Instituicao;
+import com.acolher.api.domain.Usuario;
 import com.acolher.api.repository.EnderecoRepository;
+import com.acolher.api.repository.InstituicaoRepository;
+import com.acolher.api.repository.UsuarioRepository;
 import com.acolher.api.service.EnderecoService;
 
 @Service
@@ -18,11 +23,16 @@ public class EnderecoServiceImpl implements EnderecoService{
 
 	private final Logger log = LoggerFactory.getLogger(EnderecoServiceImpl.class);
 	private final EnderecoRepository enderecoRepository;
+	private final UsuarioRepository usuarioRepository;
+	private final InstituicaoRepository instituicaoRepository;
 	
-	public EnderecoServiceImpl(EnderecoRepository enderecoRepository) {
+	public EnderecoServiceImpl(EnderecoRepository enderecoRepository, UsuarioRepository usuarioRepository,
+			InstituicaoRepository instituicaoRepository) {
 		this.enderecoRepository = enderecoRepository;
+		this.usuarioRepository = usuarioRepository;
+		this.instituicaoRepository = instituicaoRepository;
 	}
-	
+
 	@Override
 	public Optional<Endereco> getById(Integer codigo) {
 		log.debug("Endereco getById {} " , codigo);
@@ -47,6 +57,12 @@ public class EnderecoServiceImpl implements EnderecoService{
 	@Override
 	public void delete(Integer codigo) {
 		log.debug("delete endereco {} " , codigo);
+		
+		List<Usuario> usuarios = this.usuarioRepository.findAllEnderecoByCodigo(codigo);
+		List<Instituicao> instituicoes = this.instituicaoRepository.findAllEnderecoByCodigo(codigo);
+		if(usuarios.size() > 0 || instituicoes.size() > 0) {
+			throw new InvalidParameterException("Endereço não pode ser deletado pois está assossiado ao usuario");
+		}
 		
 		this.enderecoRepository.deleteById(codigo);
 	}
